requirements.txt
fastapi==0.115.1
fastapi-cli==0.0.7
uvicorn==0.34.0
python-jose[cryptography]==3.4.0
pyjwt==2.10.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.2
redis==5.0.4
aiocache==0.12.2
pytest==8.2.2
aioredis==2.0.1
python-multipart==0.0.9
lxml==5.2.1